requirements.txt
pandas
matplotlib
plotly
scipy
jinja2
pyyaml
pytest
